Somewhat improve jerking of LERP when locking on for camera recenter
Locking onto targets would not freeze the LERP back to camera center.  So simply added a check to see if player is pressing/holding the lock on button and don't let LERP run if they are, so we still get proper LERP.

This is far from perfect, because one can still follow a pattern of locking on, aiming back up or down before center is reached, locking on again and the LERP will still be very high until it snaps /jerks again.  Would be best to reset LERP value.